chore(examples): declare the web-inspector dependency in the angular demo and storybook fixes
