+++ examples/v2/angular/demo/package.json
@@ -29,6 +29,7 @@
     "@copilotkit/core": "workspace:*",
     "@copilotkit/shared": "workspace:*",
     "@copilotkit/angular": "workspace:*",
+    "@copilotkit/web-inspector": "workspace:*",
     "@jetbrains/websandbox": "^1.1.3",
     "@lucide/angular": "^1.31.0",
     "@tanstack/pacer": "^0.20.1",
