+++ examples/v2/angular/storybook/package.json
@@ -30,15 +30,16 @@
     "@angular/compiler-cli": "22.0.0",
     "@copilotkit/typescript-config": "workspace:*",
     "@copilotkit/angular": "workspace:*",
-    "@storybook/addon-themes": "10.5.8",
-    "@storybook/angular": "10.5.8",
+    "@copilotkit/web-inspector": "workspace:*",
+    "@storybook/addon-themes": "10.6.0",
+    "@storybook/angular": "10.6.0",
     "@tailwindcss/postcss": "^4.1.12",
     "@types/node": "^22",
     "autoprefixer": "^10.4.21",
     "css-loader": "^7.1.2",
     "postcss": "^8.4.31",
     "postcss-loader": "^8.1.1",
-    "storybook": "10.5.8",
+    "storybook": "10.6.0",
     "style-loader": "^4.0.0",
     "tailwindcss": "^4.1.11",
     "typescript": "6.0.3"
