+++ examples/v2/angular/storybook/stories/CopilotChatView-Components.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/angular";
-import { moduleMetadata } from "@storybook/angular";
+import { applicationConfig, moduleMetadata } from "@storybook/angular";
 import { CommonModule } from "@angular/common";
-import { Component, Injectable, signal } from "@angular/core";
+import { Component, Injectable } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import {
   CopilotChatView,
@@ -11,31 +11,19 @@ import {
   provideCopilotChatLabels,
   provideCopilotKit,
 } from "@copilotkit/angular";
+import { StoryChatState } from "./story-chat-state";
 import type { Message } from "@ag-ui/client";
 import { CustomDisclaimerComponent } from "../components/custom-disclaimer.component";
 import { CustomInputComponent } from "../components/custom-input.component";
 import { CustomScrollButtonComponent } from "../components/custom-scroll-button.component";
 
-@Injectable()
-class StoryChatState extends ChatState {
-  readonly inputValue = signal<string>("");
-
-  submitInput(value: string): void {
-    const trimmed = value.trim();
-    if (!trimmed) return;
-    console.log("[Storybook] submitInput", trimmed);
-    this.inputValue.set("");
-  }
-
-  changeInput(value: string): void {
-    this.inputValue.set(value);
-  }
-}
-
 const meta: Meta<CopilotChatView> = {
   title: "UI/CopilotChatView/Customized with Components",
   component: CopilotChatView,
   decorators: [
+    applicationConfig({
+      providers: [provideCopilotKit()],
+    }),
     moduleMetadata({
       imports: [
         CommonModule,
@@ -44,7 +32,6 @@ const meta: Meta<CopilotChatView> = {
         CopilotChatInput,
       ],
       providers: [
-        provideCopilotKit({}),
         provideCopilotChatLabels({
           chatInputPlaceholder: "Type a message...",
           chatDisclaimerText:
