+++ examples/v2/angular/storybook/stories/CopilotChatView-Templates.stories.ts
@@ -1,13 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/angular";
-import { moduleMetadata } from "@storybook/angular";
+import { applicationConfig, moduleMetadata } from "@storybook/angular";
 import { CommonModule } from "@angular/common";
-import {
-  Component,
-  Injectable,
-  Input,
-  signal,
-  ChangeDetectionStrategy,
-} from "@angular/core";
+import { Component, Input, ChangeDetectionStrategy } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import {
   CopilotChatView,
@@ -17,29 +11,17 @@ import {
   provideCopilotChatLabels,
   provideCopilotKit,
 } from "@copilotkit/angular";
+import { StoryChatState } from "./story-chat-state";
 import type { Message } from "@ag-ui/client";
-
-@Injectable()
-class StoryChatState extends ChatState {
-  readonly inputValue = signal<string>("");
-
-  submitInput(value: string): void {
-    const trimmed = value.trim();
-    if (!trimmed) return;
-    console.log("[Storybook] submitInput", trimmed);
-    this.inputValue.set("");
-  }
-
-  changeInput(value: string): void {
-    this.inputValue.set(value);
-  }
-}
 
 // Custom input components defined after imports
 const meta: Meta<CopilotChatView> = {
   title: "UI/CopilotChatView/Customized with Templates",
   component: CopilotChatView,
   decorators: [
+    applicationConfig({
+      providers: [provideCopilotKit()],
+    }),
     moduleMetadata({
       imports: [
         CommonModule,
@@ -49,7 +31,6 @@ const meta: Meta<CopilotChatView> = {
         CopilotChatInput,
       ],
       providers: [
-        provideCopilotKit({}),
         provideCopilotChatLabels({
           chatInputPlaceholder: "Type a message...",
           chatDisclaimerText:
