+++ packages/angular/src/lib/config.ts
@@ -123,7 +123,7 @@ export function injectCopilotKitConfig(): CopilotKitConfig {
   return inject(COPILOT_KIT_CONFIG);
 }
 
-export function provideCopilotKit(config: CopilotKitConfig): Provider {
+export function provideCopilotKit(config: CopilotKitConfig = {}): Provider {
   const resolvedLicense = resolveLicense(config);
   const headers = config.headers ?? {};
   if (
